Squashed commit of the following:
commit d41073aed97610e11974b3540e3d89f12b7b1183
Author: NEP\hurch <hurchanqwer@naver.com>
Date:   Wed Apr 2 17:53:31 2025 +0900

    - 플레이어, 적 피격 시 붉은 이펙트 추가
    - 적 애니메이션 분리

commit ce9b9716a9646482e34beed853c7abba65a98b73
Author: NEP\hurch <hurchanqwer@naver.com>
Date:   Wed Apr 2 17:25:56 2025 +0900

    적 공격 이펙트추가

commit eb26534d214e0282870aa3402b500f57fd6352e8
Author: NEP\hurch <hurchanqwer@naver.com>
Date:   Wed Apr 2 16:50:34 2025 +0900

    플레이어 적 일반공격 구현

commit 7ff69ddfff25e55676126fb5bb713fb1478b7146
Author: NEP\hurch <hurchanqwer@naver.com>
Date:   Wed Apr 2 10:25:31 2025 +0900

    적, 스킬 시스템 머지 완료

commit 5999d388befcda6faaa03362d9d39e1c7cd0e240
Merge: 6d33352 7bd3a02
Author: NEP\hurch <hurchanqwer@naver.com>
Date:   Wed Apr 2 10:14:11 2025 +0900

    Merge remote-tracking branch 'origin/SkillPrototype' into EnemyState

    # Conflicts:
    #	Assets/Scenes/Main.unity
    #	Assets/Scripts/Player/Player.cs

commit 6d3335248e29282bc2091faef52f5ed0ab2a564b
Author: DESKTOP-KU2CFGR\hurch <hurchanqwer>
Date:   Wed Apr 2 00:57:41 2025 +0900

    공격이펙트 애셋 추가 / 플레이어 공격 및 패링 시 넉백

commit ec54464a869ce96540bdd5e81ed59fcafba48f9c
Author: DESKTOP-KU2CFGR\hurch <hurchanqwer>
Date:   Tue Apr 1 15:36:56 2025 +0900

    패링 코루틴 null 체크

commit d991e8d85f869bd14dcec999a49b877b87185ab3
Author: DESKTOP-KU2CFGR\hurch <hurchanqwer>
Date:   Tue Apr 1 14:46:50 2025 +0900

    merge Complete

commit be8c7589e97312088b490b2fd337b2136f687042
Merge: edec42f 1ce6ac1
Author: DESKTOP-KU2CFGR\hurch <hurchanqwer>
Date:   Tue Apr 1 14:34:39 2025 +0900

    Merge branch 'Merge' into SkillPrototype

commit 1ce6ac1be316b19bfc790c4674653e9a86eae847
Author: DESKTOP-KU2CFGR\hurch <hurchanqwer>
Date:   Tue Apr 1 14:22:37 2025 +0900

    적 칼 패턴 추가
    todo. 이펙트 넣어야합니다.

+++ Assets/Scripts/EnemyInfo/EnemyData.cs
@@ -12,13 +12,12 @@ public enum EEnemyType{
 [CreateAssetMenu(fileName = "EnemyData", menuName = "Enemy/EnemyData")]
 public class EnemyData : ScriptableObject
 {
-    
     public EEnemyType eEnemyType;
     public string monsterName;
     public int maxHealth;
     public float moveSpeed;
     public float attackSpeed;
-    public float attackDamage;
+    public int attackDamage;
